% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bst235Project-package.R
\docType{package}
\name{bst235Project-package}
\alias{bst235Project}
\alias{bst235Project-package}
\title{bst235Project: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
