% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bst235Project-package.R
\docType{package}
\name{bst235Project}
\alias{bst235Project}
\alias{bst235Project-package}
\title{bst235Project: BIOSTAT 235 course project, Fall 2017.}
\description{
This package contains the code for the BIOSTAT 235 Fall 2017 course
project, problem 5.
}
