% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_true_model.R
\name{fit_true_model}
\alias{fit_true_model}
\alias{objective}
\alias{gradient}
\title{Fit true regression model}
\usage{
fit_true_model(xs, ys)

objective(alpha_betas, xs, ys)

gradient(alpha_betas, xs, ys)
}
\arguments{
\item{xs}{n x p matrix of predictors}

\item{ys}{Length-n vector of outcomes}

\item{alpha_betas}{Value of coefficients to evaluate the function.
First entry must be the intercept, and the rest are the other
coefficients.}
}
\value{
\code{fit_true_model} returns a length-(p + 1) vector of
coefficients, where the first is the intercept and the rest are the
beta values. If the fitting algorithm did not converge,
\code{fit_true_model} will fail with an error.

\code{objective} and \code{gradient} return the value of the
objective function and gradient evaluated for some coefficients
alpha_beta, respectively.
}
\description{
Fit true regression model E(Y_i | X_i) = alpha_0 + Phi(beta^T X_i),
where Phi is the standard Normal CDF.
}
\details{
Under the hood, the fitting is done using BFGS, as implemented in the
\code{optim} function in R.
}
\section{Functions}{
\itemize{
\item \code{objective}: Evaluate objective function

\item \code{gradient}: Evaluate objective function's gradient
}}

