% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{link_viol_sim}
\alias{link_viol_sim}
\title{Simulate violations of link function}
\usage{
link_viol_sim(nsims, beta, x_simulator, true_link, wrong_link,
  error_simulator = rnorm, ns = c(10, 100), lambdas = ns^(1/3),
  kernel = dnorm, bandwidths = c(0.1, 0.05))
}
\arguments{
\item{nsims}{Number of simulations to run.}

\item{beta}{True effects. Should be a length-p vector.}

\item{x_simulator}{Function whose first argument is n. Generates n
replicates of X. The return value of this function should be an n x
p matrix, or an n x 1 vector.}

\item{true_link}{True link function. Should be an R function that
takes one argument.}

\item{wrong_link}{Misspecified link function. Should be an R function
that takes one argument.}

\item{error_simulator}{Function whose first argument is n. Generates
n replicates of epsilon. The return value of this function should
be an n x 1 vector.}

\item{ns}{Vector of n to use.}

\item{lambdas}{Vector of lambdas to use. Should be the same length as
ns.}

\item{kernel}{Kernel function to use for smoothing the predicted
values.}

\item{bandwidths}{Vector of bandwidths to use for smoothing the
predicted values.}
}
\value{
A list with two named elements
\describe{
\item{betas}{nsims x p DataFrame of estimated coefficients for each
iteration of the simulation.}
\item{errors}{nsims x 3 DataFrame of three estimated errors.}
}
}
\description{
Simulate violations of link function
}
\examples{
1 + 1
}
