% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{np_calibrate}
\alias{np_calibrate}
\title{Nonparametric calibration step for predicting}
\usage{
np_calibrate(new_xs, xs, ys, glmnet_obj, bandwidth)
}
\arguments{
\item{new_xs}{Matrix of new predictors}

\item{xs}{Matrix of old predictors}

\item{ys}{Vector of old observed values}

\item{glmnet_obj}{A glmnet object for which there is a predict method}
}
\value{
Nonparametrically calibrated predicted means.
}
\description{
Nonparametric calibration step for predicting the mean
}
\details{
The formula in the assignment does not include intercepts, but we do
include the intercepts here by using the \code{predict} function.
However, because the intercepts will be the same, after subtraction
they cancel out.
}
